% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beautify_arm_var}
\alias{beautify_arm_var}
\title{Give arm variable proper form and names}
\usage{
beautify_arm_var(arm)
}
\arguments{
\item{arm}{character vector, will be converted to factor and renamed}
}
\value{
A factor with the right labels in the right order for pretty printing and plotting
}
\description{
Give arm variable proper form and names
}
\keyword{internal}
