% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_mean_diffs}
\alias{bootstrap_mean_diffs}
\title{Bootstrap mean differences}
\usage{
bootstrap_mean_diffs(vals, grps, B = 2000L, fixed_seed = 4131L)
}
\arguments{
\item{vals}{numeric vector of values from observations}

\item{grps}{int vector with group assignments of the values in the `vals` parameter.
Assumes: 0 = actv, 1 = ctrl.}

\item{B}{int, the number of bootstrap samples (default = 2000)}

\item{fixed_seed}{int, seed to be used for pseudo-random number generation}
}
\value{
bootstrap samples of the mean difference by group assignment, as numeric vector.
}
\description{
Bootstrap mean differences
}
