% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimates}
\alias{compute_estimates}
\title{Compute outcome estimates for a given patient}
\usage{
compute_estimates(
  t_icu_discharge,
  t_death,
  first_hrqol_patient,
  is_mortality_benefitter,
  acceleration_hrqol,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  mortality_trajectory_shape,
  mortality_dampening,
  sampling_frequency
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{first_hrqol_patient}{scalar, the HRQoL value at ICU discharge of this patient}

\item{is_mortality_benefitter}{logical, is patient a mortality benefitters? Defaults to FALSE.}

\item{acceleration_hrqol}{scalar in \verb{[0, Inf)}, acceleration of HRQoL improvement compared to baseline
trajectory}

\item{first_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{sampling_frequency}{int (scalar or named vector), span between HRQoL sampling from patients}
}
\value{
A six-element named numeric vector with single-sampled HRQoL at end of follow-up and area
under the HRQoL trajectory for the primary and the two secondary analyses.
}
\description{
Compute outcome estimates for a given patient
}
\keyword{internal}
