% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_performance_metrics}
\alias{compute_performance_metrics}
\title{Computes select Monte Carlo performance measures}
\usage{
compute_performance_metrics(
  theta_hat,
  theta,
  p_value,
  ci_lo,
  ci_hi,
  alpha = 0.05
)
}
\arguments{
\item{theta_hat}{vector, estimates}

\item{theta}{vector, ground truth values (length 1 or as long as `theta_hat`)}

\item{p_value, ci_lo, ci_hi}{vectors holding the p values as well as lower and upper boundaries of
the confidence intervals corresponding to the alpha (next parameter)}

\item{alpha}{scalar in [0, 1], equals 1.0 - significance level}
}
\value{
A named numeric vector with the performance measures.
}
\description{
Computes select Monte Carlo performance measures
}
\keyword{internal}
