% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_arm_level_trajectory.R
\name{construct_arm_level_trajectory}
\alias{construct_arm_level_trajectory}
\title{Title}
\usage{
construct_arm_level_trajectory(
  t_icu_discharge,
  sampling_frequency = 14L,
  acceleration_hrqol = 0,
  start_hrqol = 0.1,
  final_hrqol = 0.75
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{sampling_frequency}{int, for span between samplings from patients}

\item{start_hrqol}{scalar, HRQoL at ICU discharge}

\item{mortality_dampening}{scalar, dampening effect of HRQoL at ICU discharge in patients who die before end of follow-up}

\item{acceleration_hrqol_actv}{scalar,}

\item{min_hrqol}{sclar HRQoL at end of follow-up}
}
\value{
A matrix with two columns: x is time and y the corresponding HRQoL values.
}
\description{
Uses dynamic caching (powered by memoise package) for increased performance. Default setting correspond to control-arm settings.
}
\keyword{internal}
