% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_final_trajectories.R
\name{construct_final_trajectories}
\alias{construct_final_trajectories}
\title{Compute final trajectories}
\usage{
construct_final_trajectories(traj, t_icu_discharge, sampling_frequency)
}
\arguments{
\item{traj}{matrix, x = time and y = HRQoL}

\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{sampling_frequency}{int (scalar or named vector), span between HRQoL sampling from patients}
}
\value{
A three-element list with trajectories for each of the three assumptions as to when
sampling can start.
}
\description{
Cached. Takes a "raw" trajectory object and samples values according to sampling strategy.
Assumes the first coordinate in the \code{traj} object is day of ICU discharge.
}
