% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_mortality_trajectory.R
\name{construct_mortality_trajectory}
\alias{construct_mortality_trajectory}
\title{Construct mortality trajectory for a patient}
\usage{
construct_mortality_trajectory(
  t_death,
  t_icu_discharge,
  start_hrqol = 0,
  mortality_trajectory_shape = "exp_decay",
  resolution = 100,
  n_halflives_to_zero = 10
)
}
\arguments{
\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{start_hrqol}{scalar, HRQoL at ICU discharge}

\item{mortality_trajectory_shape}{string, should be any of the following four: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}}

\item{resolution}{int, the number of grid points on the HRQoL trajectory between ICU discharge
and time of death}

\item{n_halflives_to_zero}{scalar, the number of half-lives after which one considers HRQoL to
equal \code{0}.}
}
\value{
Two-column matrix: x is time, y is the HRQoL.
}
\description{
If a patient dies in the follow-up period, they will follow the HRQoL trajectory produced by this
function.
}
\keyword{internal}
