% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_patient_trajectory.R
\name{construct_patient_trajectory}
\alias{construct_patient_trajectory}
\title{Title}
\usage{
construct_patient_trajectory(
  t_icu_discharge = NULL,
  sampling_frequency = 14L,
  acceleration_hrqol = 0,
  start_hrqol_arm = 0.1,
  start_hrqol_patient = start_hrqol_arm,
  final_hrqol_arm = 0.75,
  t_death = Inf,
  is_mortality_benefitter = FALSE,
  mortality_trajectory = "exp_decay",
  mortality_dampening = 0,
  n_digits = 3
)
}
\arguments{
\item{n_digits}{int, the number of digits of HRQoL values}
}
\description{
Title
}
