% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crayon_style}
\alias{crayon_style}
\title{Colour text if crayon package is available}
\usage{
crayon_style(x, style)
}
\arguments{
\item{x}{string to be coloured}

\item{style}{string, must be a valid \code{crayon} style}
}
\value{
Coloured string (if crayon available), otherwise the string remains unaltered
}
\description{
Colour text if crayon package is available
}
\keyword{internal}
