% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_smooth_trajectory.R
\name{create_smooth_trajectory}
\alias{create_smooth_trajectory}
\title{Create smooth trajectory}
\usage{
create_smooth_trajectory(x, y, alpha = 0.5, epsilon = 1e-04)
}
\arguments{
\item{x, y}{numeric vectors with points to be connected with smooth curve}

\item{alpha}{scalar in \verb{[0, 1]}, controls the tightness of the curve. \code{alpha = 0.5} (the default)
yields a centripetal Catmull-Rom curve}

\item{epsilon}{small scalar, used to create phantom points. The default value should suffice for
most use cases (see also link below).}
}
\value{
A data.table with the (x, y) coordinates of the smooth trajectory.
}
\description{
Default settings create a centripetal Catmull-Rom curve with the coordinates given by \code{x} and
\code{y}.
}
\details{
This function is based on the nice run-down of smooth curves at
https://apoorvaj.io/cubic-bezier-through-four-points.
}
