% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_xout}
\alias{create_xout}
\title{Create xout vector for approx()}
\usage{
create_xout(start, end, by)
}
\arguments{
\item{by}{number: increment of the sequence.}
}
\description{
Cached. Used in \code{construct_final_trajectories()}.
}
\keyword{internal}
