% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_remaining_time}
\alias{est_remaining_time}
\title{Estimate remaining time}
\usage{
est_remaining_time(start_time, i, n, now = Sys.time())
}
\arguments{
\item{start_time}{object of time \code{"POSIXct"} (e.g. output from \code{Sys.time()})}

\item{i}{single integer, current iteration}

\item{n}{single integer, total number iteration to go through}
}
\value{
Character string with estimate
}
\description{
Uses a counter and the total iterations to estimate how much time remains
}
\keyword{internal}
