% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helper.R
\name{estimation_helper}
\alias{estimation_helper}
\title{Helper function to estimate outcomes in individual patients}
\usage{
estimation_helper(
  n_patients,
  arm,
  prop_mortality_benefitters_actv,
  start_hrqol_arm,
  inter_patient_noise_sd,
  acceleration_hrqol,
  mortality_trajectory_shape,
  mortality_dampening,
  mortality_rng,
  n_digits = 2,
  min_valid_hrqol = -0.757
)
}
\arguments{
\item{n_patients}{int, the number of patients for whom to estimate outcomes}

\item{arm}{string, either \code{"actv"} or \code{"ctrl"}}

\item{prop_mortality_benefitters_actv}{scalar numeric in \verb{[0, 1]}, the proportion of patients in
the active arm who are so-called mortality benefitters.}

\item{start_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{acceleration_hrqol}{scalar, acceleration of HRQoL improvement}

\item{mortality_trajectory_shape}{string, should be any of the following four: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}}

\item{mortality_dampening}{scalar, dampening effect of HRQoL at ICU discharge in patients who die
before end of follow-up}

\item{mortality_rng}{function, a function that produces times of death (only one parameter which should be the number of values to sample)}

\item{n_digits}{int, the number of digits of HRQoL values}

\item{min_valid_hrqol}{scalar, the minimum valid HRQoL value}
}
\value{
Data.table with outcome results for all patients. NB! There are no trial_id indicator as
this is not required in all uses of this helper function.
}
\description{
Helper function to estimate outcomes in individual patients
}
\keyword{internal}
