% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_approx}
\alias{fast_approx}
\title{Fast approximation function}
\usage{
fast_approx(x, y, xout, method = 1, na.rm = TRUE)
}
\arguments{
\item{x, y}{numeric vectors giving the coordinates of the points to be
    interpolated.  Alternatively a single plotting structure can be
    specified: see \code{\link{xy.coords}}.}

\item{xout}{an optional set of numeric values specifying where
    interpolation is to take place.}

\item{method}{int, 1 = linear, 2 = constant}

\item{na.rm}{logical specifying how missing values (\code{\link{NA}}'s)
    should be handled.  Setting \code{na.rm=FALSE} will propagate
    \code{NA}'s in \code{y} to the interpolated values, also depending on
    the \code{rule} set.  Note that in this case, \code{NA}'s in \code{x}
    are invalid, see also the examples.}
}
\value{
interpolated value(s) corresponding to `xout`
}
\description{
This function essentially strips unnecessary housekeeping from
stats::approx() to yield some 30x speed-up. Note that this function call
non-exported code from the stats package.
}
\keyword{internal}
