% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_auc}
\alias{fast_auc}
\title{Area under the curve}
\usage{
fast_auc(x, y)
}
\arguments{
\item{x, y}{numeric vectors of coordinates}
}
\value{
Area under the curve as a scalar.
}
\description{
Area under the curve
}
