% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_decay_halflife.R
\name{find_decay_halflife}
\alias{find_decay_halflife}
\title{Find appropriate half-life of exponential decay}
\usage{
find_decay_halflife(
  t_diff = 180 - 1,
  t_start = 0,
  t_end = 180,
  res = 0.1,
  tol = 0.001
)
}
\arguments{
\item{tol}{}
}
\description{
The notion of this function is to find the largest half-life that yields a almost-zero decay factor (as per `tol` argument) at `grid_end`.
}
