% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mortality_funs.R
\name{generate_mortality_funs}
\alias{generate_mortality_funs}
\title{Generate density, probability, quantile and random generation functions for
(scaled) empirical mortality distribution}
\usage{
generate_mortality_funs(
  cum_mortality = NULL,
  censoring_value = 181,
  censoring_threshold = NULL
)
}
\arguments{
\item{censoring_threshold}{}
}
\description{
Generate density, probability, quantile and random generation functions for
(scaled) empirical mortality distribution
}
