% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_approx}
\alias{linear_approx}
\title{Interpolate multiple values}
\usage{
linear_approx(x, y, xout)
}
\arguments{
\item{x, y}{coordinates for known observationa}

\item{xout}{vector or scalar with value(s) where interpolation should happen}
}
\value{
Interpolated values
}
\description{
User-facing, stripped-down version of approx.c from the stats package (which is around twice
as fast as this one).
Yields somewhere around 10x speed-up compared to using stats::approx().
}
