% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_timediff}
\alias{log_timediff}
\title{Helper for printing time difference to stdout as message}
\usage{
log_timediff(start_time, msg = NULL, style = NULL, now = Sys.time())
}
\arguments{
\item{start_time, now}{output from \code{Sys.time()}}

\item{msg}{str, message to print}
}
\description{
Helper for printing time difference to stdout as message
}
\keyword{internal}
