% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{measure_memory_use}
\alias{measure_memory_use}
\title{Helper to track memory use (if pryr package available)}
\usage{
measure_memory_use(current_max)
}
\arguments{
\item{current_max}{\code{NULL}, object of class \code{pryr_bytes} or scalar with bytes.}
}
\value{
Numeric scalar of class \code{pryr_bytes}.
}
\description{
Helper to track memory use (if pryr package available)
}
\keyword{internal}
