% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OR}
\alias{OR}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{arbitrary vectors}
}
\description{
Infix function for easy replacement of NULL with a default value.
}
\keyword{internal}
