% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad}
\alias{pad}
\title{Padding of string}
\usage{
pad(x, n, pad = " ", side = "right")
}
\arguments{
\item{x}{string, to be padded}

\item{n}{int, number of padding characters to add}

\item{pad}{string, padding character(s)}

\item{side}{string, where to pad \code{x} (default: \code{"right"})}
}
\value{
Padded character
}
\description{
Padding of string
}
\keyword{internal}
