% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_examples.R
\name{plot_results}
\alias{plot_results}
\alias{plot_examples.hrqolr_results}
\title{Plot example trajectories}
\usage{
plot_results(object, ...)

plot_examples.hrqolr_results(object, ...)
}
\arguments{
\item{object}{object of class \code{hrqolr_results} as returned by \code{simulate_trial}}

\item{...}{passed on to methods}
}
\description{
Plot example trajectories

Actual function
}
