% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.hrqolr_examples}
\alias{print.hrqolr_scenario_validation_results}
\alias{print.hrqolr_scenario}
\alias{print.hrqolr_comparisons}
\alias{print.hrqolr_summary_stats}
\alias{print.hrqolr_bytes}
\title{Print methods for HRQoL results}
\usage{
\method{print}{hrqolr_examples}(x, prop_digits = 3, ...)

\method{print}{hrqolr_scenario_validation_results}(x, ...)

\method{print}{hrqolr_scenario}(x, ...)

\method{print}{hrqolr_comparisons}(x, n_digits = 3, ...)

\method{print}{hrqolr_summary_stats}(x, n_digits = 3, ...)

\method{print}{hrqolr_bytes}(x, digits = 3, ...)
}
\arguments{
\item{x}{object to print, see \strong{Details}.}

\item{prop_digits}{single integer (default is \code{3}), the number of digits used
when printing proportions.}

\item{...}{passed on to methods}

\item{n_digits}{int, number of digits to show in decimal numbers}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints contents of the first input \code{x} in a human-friendly way, see
\strong{Details} for more information.

Logic from pryr:::print.pryr_bytes but simplified and included here to mimise dependencies.
}
\details{
The behaviour depends on the class of \code{x}:
}
