% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_infinite}
\alias{replace_infinite}
\alias{\%fi\%}
\title{Default value for infinities}
\usage{
a \%fi\% b
}
\arguments{
\item{a, b}{arbitrary vectors}
}
\description{
Infix function for easy replacement of -Inf/Inf with a default value.
}
\keyword{internal}
