% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na}
\alias{replace_na}
\title{Default value for NA's}
\usage{
replace_na(x, replacement)
}
\arguments{
\item{x}{arbitrary vector}

\item{replacement}{replacement value}
}
\value{
x vector in which NA's have been replaced with `replacement`.
}
\description{
Helper function for simple handling of missing values in vectors. No type checking done.
}
\keyword{internal}
