% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Scale numeric vector to sum to 1}
\usage{
rescale(x)
}
\arguments{
\item{x}{numeric vector, should not contain NA's}
}
\description{
Scale numeric vector to sum to 1
}
\keyword{internal}
