% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_example_trajectories.R
\name{sample_example_trajectories}
\alias{sample_example_trajectories}
\alias{sample_example_trajectories.hrqolr_scenario}
\alias{sample_example_trajectories.default}
\title{Sample example trajectories}
\usage{
sample_example_trajectories(scenario, ...)

\method{sample_example_trajectories}{hrqolr_scenario}(
  scenario,
  n_example_trajectories_per_arm = 50,
  n_digits = 2,
  seed = NULL,
  valid_hrqol_range = c(-0.757, 1),
  ...
)

\method{sample_example_trajectories}{default}(
  arms,
  sampling_frequency,
  index_hrqol,
  first_hrqol,
  final_hrqol,
  acceleration_hrqol,
  mortality,
  prop_mortality_benefitters,
  mortality_trajectory_shape,
  mortality_dampening,
  n_example_trajectories_per_arm,
  n_digits,
  seed = NULL,
  valid_hrqol_range,
  ...
)
}
\arguments{
\item{scenario}{object of class 'hrqolr_scenario', the output of \link{setup_scenario}}

\item{...}{passed on to methods}

\item{n_example_trajectories_per_arm}{int, the number of example trajectories to include in the
returned object}

\item{n_digits}{int, the number of digits of HRQoL values. More digits will yield greater
precision but also cause longer run-times.}

\item{seed}{int, optional seed for reproducible pseudo-random number generation. Defaults to a
deterministic value based on the arguments given (ensuring reproducibility by default).}

\item{valid_hrqol_range}{two-element numeric vector, the lower and upper bounds of valid HRQoL
values. The default (\code{c(-0.757, 1.0)}) corresponds to the Danish EQ-5D-5L index values.}
}
\value{
An object of class \code{hrqolr_trajectories} with trajectories in data.tables. The function will
always return arm-level trajectories. If \code{n_example_trajectories_per_arm > 0}, it will also return
data.table with patient trajectories, with a \code{id} column denoting patients.
}
\description{
Sample a desired number of trajectories at arm leven and for individual patients.

Internal function that shouldn't really be invoked by the user directly. The arguments given must
be named vectors.
}
\keyword{internal}
