% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenario.R
\name{setup_scenario}
\alias{setup_scenario}
\title{Helper for setting up valid scenarios}
\usage{
setup_scenario(
  arms = NULL,
  n_patients = NULL,
  index_hrqol = NULL,
  first_hrqol = NULL,
  final_hrqol = NULL,
  acceleration_hrqol = NULL,
  mortality = NULL,
  mortality_dampening = NULL,
  mortality_trajectory_shape = NULL,
  prop_mortality_benefitters = NULL,
  sampling_frequency = NULL,
  verbose = TRUE
)
}
\arguments{
\item{arms}{character vector with the names of the arms. Must match the names of named vectors
below.}

\item{n_patients}{named int vector, number of patient in each arm}

\item{index_hrqol}{named numeric vector, the HRQoL at index (= enrolment)}

\item{first_hrqol}{named numeric vector, the HRQoL at ICU discharge in each arm}

\item{final_hrqol}{named numeric vector, the HRQoL at end of follow-up in each arm}

\item{acceleration_hrqol}{named numeric vector, relative acceleration of HRQoL improvement in
each arm}

\item{mortality}{named numeric vector in \verb{[0, 1]}, the mortality in at end of follow-up, in each
arm}

\item{mortality_dampening}{named numeric vector, dampening effect on HRQoL at ICU discharge in
patients who die before end of follow-up}

\item{mortality_trajectory_shape}{named character vector, valid values are: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}. Can differ across arms.}

\item{prop_mortality_benefitters}{named numeric vector \verb{[0, 1]}, the proportion of patients in
each arm who are so-called mortality benefitters.}

\item{sampling_frequency}{named int vector, span between HRQoL sampling from patients, in each arm.}

\item{verbose}{logical, should the function provide detailed output on validation results?
Default: \code{TRUE}}
}
\value{
An object of class 'hrqolr_scenario' with validated inputs.
}
\description{
Inputs are validated to ensure they produce the expected results when given as the input in
\link{simulate_trials}. Therefore, any arguments to this function can also be just a single value
(except the \code{arms} argument), and this function will expand such single-valued arguments to be
compatible with \code{simulate_trials}.
}
