% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_trial.default}
\alias{simulate_trial.default}
\title{Internal default workhorse function for simulate_trial}
\usage{
\method{simulate_trial}{default}(
  arms,
  n_patients,
  sampling_frequency,
  index_hrqol,
  first_hrqol,
  final_hrqol,
  acceleration_hrqol,
  mortality,
  mortality_dampening,
  mortality_trajectory_shape,
  prop_mortality_benefitters,
  test_fun,
  verbose = FALSE,
  n_digits,
  valid_hrqol_range,
  alpha,
  seed = NULL,
  ...
)
}
\arguments{
\item{arms}{character vector with the names of the arms}

\item{n_patients}{named integer (single value or named vector), number of patients}

\item{sampling_frequency}{integer (single value or named vector), span between HRQoL sampling from patients}

\item{index_hrqol}{numeric (single value or named vector), the HRQoL at index (= enrolment)}

\item{first_hrqol}{numeric (single value or named vector), the HRQoL at ICU discharge in each arm}

\item{final_hrqol}{numeric (single value or named vector), the HRQoL at end of follow-up in each arm}

\item{acceleration_hrqol}{numeric (single value or named vector), relative acceleration of HRQoL
improvement in each arm}

\item{mortality}{numeric (single value or named vector) in \verb{[0, 1]}, the mortality in at end of
follow-up, in each arm}

\item{mortality_dampening}{numeric (single value or named vector), dampening effect on HRQoL at ICU
discharge in patients who die before end of follow-up}

\item{mortality_trajectory_shape}{character (single value or named vector), valid values are
\code{"exp_decay"} (default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"} and can differ
across arms.}

\item{prop_mortality_benefitters}{numeric (single value or named vector) \verb{[0, 1]}, the proportion of
patients in each arm who are so-called mortality benefitters}

\item{verbose}{single logical, should the function provide detailed output on validation results?
Default: \code{TRUE}}

\item{...}{not used}
}
\description{
Internal function that shouldn't really be invoked by the user directly. The arguments given must
be named vectors.
}
\keyword{internal}
