% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trials.R
\name{simulate_trials.default}
\alias{simulate_trials.default}
\title{Workhorse}
\usage{
\method{simulate_trials}{default}(
  arms,
  n_patients,
  sampling_frequency,
  index_hrqol,
  first_hrqol,
  final_hrqol,
  acceleration_hrqol,
  mortality,
  mortality_dampening,
  mortality_trajectory_shape,
  prop_mortality_benefitters,
  n_trials = 100,
  n_patients_ground_truth = 1000,
  n_example_trajectories_per_arm = 50,
  include_trial_results,
  test_fun,
  verbose,
  n_digits,
  valid_hrqol_range,
  alpha,
  max_batch_size,
  seed = NULL,
  ...
)
}
\arguments{
\item{arms}{character vector with the names of the arms}

\item{n_patients}{named int (scalar or named vector), number of patients}

\item{sampling_frequency}{int (scalar or named vector), span between HRQoL sampling from patients}

\item{index_hrqol}{numeric (scalar or named vector), the HRQoL at index (= enrolment)}

\item{first_hrqol}{numeric (scalar or named vector), the HRQoL at ICU discharge in each arm}

\item{final_hrqol}{numeric (scalar or named vector), the HRQoL at end of follow-up in each arm}

\item{acceleration_hrqol}{numeric (scalar or named vector), relative acceleration of HRQoL
improvement in each arm}

\item{mortality}{numeric (scalar or named vector) in [0, 1], the mortality in at end of
follow-up, in each arm}

\item{mortality_dampening}{numeric (scalar or named vector), dampening effect on HRQoL at ICU
discharge in patients who die before end of follow-up}

\item{mortality_trajectory_shape}{character (scalar or named vector), valid values are
\code{"exp_decay"} (default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"} and can differ
across arms.}

\item{prop_mortality_benefitters}{numeric (scalar or named vector) \verb{[0, 1]}, the proportion of
patients in each arm who are so-called mortality benefitters.}

\item{verbose}{logical, should the function provide detailed output on validation results?
Default: \code{TRUE}}

\item{...}{not used}
}
\description{
Internal function that shouldn't really be invoked by the user directly. The arguments given must
be named vectors.
}
\keyword{internal}
