% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trials.R
\name{simulate_trials}
\alias{simulate_trials}
\alias{simulate_trials.hrqolr_scenario}
\title{Simulate trials}
\usage{
simulate_trials(scenario, ...)

\method{simulate_trials}{hrqolr_scenario}(
  scenario,
  n_trials = 100,
  n_patients_ground_truth = 1000,
  n_example_trajectories_per_arm = 50,
  test_fun = welch_t_test,
  include_trial_results = FALSE,
  verbose = TRUE,
  n_digits = 2,
  valid_hrqol_range = c(-0.757, 1),
  alpha = 0.05,
  max_batch_size = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{scenario}{object of class 'hrqolr_scenario', the output of \link{setup_scenario}}

\item{...}{passed on to methods}

\item{n_trials}{int scalar or vector. If vector, simulations will be run in batches of size given}

\item{n_patients_ground_truth}{int, how many patients (per arm) to use when estimating the ground
truth}

\item{n_example_trajectories_per_arm}{int, the number of example trajectories to include in the
returned object}

\item{test_fun}{function used to compare estimates. The default is \code{welch_t_test} (built into
\code{hrqolr}). Note that the function should handle the number of arms provided. See \strong{Details}
below.}

\item{include_trial_results}{logical, indicates whether trial-level results are kept. Default is
\code{FALSE} because the resulting object may be very large if many trials are simulated.}

\item{verbose}{logical, should the function give progress timestamped updates? Default: \code{TRUE}}

\item{n_digits}{int, the number of digits of HRQoL values. More digits will yield greater
precision but also cause longer run-times.}

\item{valid_hrqol_range}{two-element numeric vector, the lower and upper bounds of valid HRQoL
values. The default (\code{c(-0.757, 1.0)}) corresponds to the Danish EQ-5D-5L index values.}

\item{alpha}{scalar in \verb{[0, 1]}, the desired type 1 error rate used when comparing HRQoL in the
arms.}

\item{max_batch_size}{int, the maximum number of patients to process in each batch. The default
is to use run one batch (i.e. no upper limit)}

\item{seed}{int, optional seed for reproducible pseudo-random number generation. Defaults to a
deterministic value based on the arguments given (ensuring reproducibility by default).}
}
\value{
An object of class \code{hrqolr_results}, which is a specialised list with four elements:
summary statistics for each arm, comparisons (incl. performance metrics), the seed and the
elapsed time.
}
\description{
This is the key user-facing function for simulating trials.
}
\details{
\itemize{
\item \code{test_fun}: [pending]
}
}
