% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_var}
\alias{summarise_var}
\title{Summarise a variable with common statistics}
\usage{
summarise_var(x, probs = c(0.25, 0.5, 0.75), na_rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{probs}{probabilities for which to compute the quantiles}

\item{na_rm}{logical, should NA's be removed (default: \code{TRUE})}
}
\description{
Input e.g. a scalar to get the corresponding names
}
\keyword{internal}
