% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{welch_t_test}
\alias{welch_t_test}
\title{Welch's t test}
\usage{
welch_t_test(vals, grps, na_replacement = NULL, alpha = 0.05)
}
\arguments{
\item{vals}{numeric vector}

\item{alpha}{numeric, 1.0 - confidence_level}

\item{grp}{character vector, allowed values are "ctrl" and "actv" and used to
slice \code{x} into two vectors with values to be compared.}
}
\value{
A four-element vector with point estimate, p value and confidence internal
}
\description{
Custom and stripped-down version of R's built-in t.test that yields the
estimate with the confidence interval corresponding to the specified alpha
level.
}
\keyword{internal}
