% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auc}
\alias{auc}
\title{Compute area under the curve}
\usage{
auc(x, y)
}
\arguments{
\item{x, y}{vectors, coordinates of the curve under which the area is to be
computed}
}
\value{
scalar
}
\description{
Assumes x and y are the same length and of type \code{double}. Checking would
cause huge performance hit and, so, is foregone.
}
\keyword{internal}
