% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_estimates.R
\name{bootstrap_estimates}
\alias{bootstrap_estimates}
\title{Bootstrap confidence intervals and p values}
\usage{
bootstrap_estimates(
  vals,
  grps,
  na_replacement = NULL,
  n_samples = 2000,
  alpha = 0.05
)
}
\arguments{
\item{vals}{numeric vector}

\item{grps}{character vector, allowed values are "ctrl" and "actv" and used to slice \code{x} into two
vectors with values to be compared.}

\item{na_replacement}{scalar (or \code{NULL}), if \code{NULL} missing values will be removed. Otherwise,
missing values will be replaced by the value of \code{na_replacement}}

\item{n_samples}{scalar, number of bootstrap samples}

\item{alpha}{numeric, 1.0 - confidence_level}
}
\value{
A four-element vector with point estimate, p value and confidence internal
}
\description{
Bootstrap confidence intervals and p values
}
\keyword{internal}
