% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_hrqolr.R
\name{cache_hrqolr}
\alias{cache_hrqolr}
\alias{clear_hrqolr_cache}
\title{Caching of hrqolr functions}
\usage{
cache_hrqolr(max_size = 2 * 1024^3)

clear_hrqolr_cache()
}
\arguments{
\item{max_size}{Maximum size of the cache, in bytes. If the cache exceeds
this size, the entire cache is wiped; this is much faster than using an
eviction policy for how the cache is used in \code{hrqolr} objects will be
removed according to the value of the}
}
\value{
\code{NULL}, these functions are called for their side effects: caching (or clearing caches) of
select \code{hrqolr} functions.
}
\description{
Caching of hrqolr functions
}
