% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_eof}
\alias{compute_eof}
\title{Compute end of follow-up}
\usage{
compute_eof(t_icu_discharge, approx_end = 180, sampling_frequency = 14)
}
\arguments{
\item{t_icu_discharge}{int, day of ICU discharge}

\item{approx_end}{int, the approximate end day (follow-up will not be shorter than this)}

\item{sampling_frequency}{int, period between samplings of HRQoL}
}
\description{
Cached. Based entirely on day of ICU discharge and sampling frequency
}
