% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimates.R
\name{compute_estimates}
\alias{compute_estimates}
\title{Compute outcome estimates for a given patient}
\usage{
compute_estimates(
  t_icu_discharge = NULL,
  t_death = NA,
  start_hrqol_patient = start_hrqol_arm,
  is_mortality_benefitter = FALSE,
  acceleration_hrqol = 0,
  start_hrqol_arm = 0.1,
  final_hrqol_arm = 0.75,
  mortality_trajectory_shape = "exp_decay",
  mortality_dampening = 0,
  sampling_frequency = 14L,
  n_digits = 2
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{start_hrqol_patient}{scalar, the HRQoL value at ICU discharge of this patient}

\item{is_mortality_benefitter}{logical, is patient a mortality benefitters? Defaults to FALSE.}

\item{start_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{mortality_trajectory_shape}{string, should be any of the following four: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}}

\item{mortality_dampening}{scalar, dampening effect of HRQoL at ICU discharge in patients who die
before end of follow-up}

\item{sampling_frequency}{int, for span between samplings from patients}

\item{n_digits}{int, the number of digits of HRQoL values}
}
\value{
A six-element named numeric vector with single-sampled HRQoL at end of follow-up and area
under the HRQoL trajectory for the primary and the two secondary analyses.
}
\description{
Compute outcome estimates for a given patient
}
\keyword{internal}
