% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_arm_level_trajectory.R
\name{construct_arm_level_trajectory}
\alias{construct_arm_level_trajectory}
\title{Construct arm-level trajectory}
\usage{
construct_arm_level_trajectory(
  t_icu_discharge,
  acceleration_hrqol,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  sampling_frequency = 14L
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{acceleration_hrqol}{scalar in \verb{[0, Inf)}, acceleration of HRQoL improvement compared to baseline
trajectory}

\item{first_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{sampling_frequency}{int (scalar or named vector), span between HRQoL sampling from patients}
}
\value{
A matrix with two columns: x is time and y the corresponding HRQoL values.
}
\description{
Default arguments correspond to control-arm settings. All patient trajectories, within a given
trial, are based on this.
}
\keyword{internal}
