% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_mortality_trajectory.R
\name{construct_mortality_trajectory}
\alias{construct_mortality_trajectory}
\title{Construct mortality trajectory for a patient}
\usage{
construct_mortality_trajectory(
  t_death,
  t_icu_discharge,
  index_hrqol,
  first_hrqol,
  mortality_trajectory_shape,
  resolution = 100,
  n_halflives_to_zero = 10
)
}
\arguments{
\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{index_hrqol}{single numeric value, HRQoL at index (= time of enrolment)}

\item{first_hrqol}{single numeric value, HRQoL at ICU discharge}

\item{resolution}{single integer, the number of grid points on the HRQoL trajectory between ICU discharge
and time of death}

\item{n_halflives_to_zero}{single numeric value, the number of half-lives after which one considers HRQoL to
equal \code{0}.}
}
\value{
Two-column matrix: \code{x} is time, \code{y} is the HRQoL.
}
\description{
If a patient dies in the follow-up period, they will follow the HRQoL trajectory produced by this
function.
}
\keyword{internal}
