% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_patient_trajectory.R
\name{construct_patient_trajectory}
\alias{construct_patient_trajectory}
\title{Construct the full (primary) HRQoL trajectory for a patient}
\usage{
construct_patient_trajectory(
  t_icu_discharge,
  t_death,
  first_hrqol_patient,
  acceleration_hrqol,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  is_mortality_benefitter,
  mortality_trajectory_shape,
  mortality_dampening,
  sampling_frequency = 14L,
  n_digits = 2
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{first_hrqol_patient}{scalar, the HRQoL value at ICU discharge of this patient}

\item{acceleration_hrqol}{scalar, acceleration of HRQoL improvement compared to baseline
trajectory}

\item{first_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{is_mortality_benefitter}{logical, is patient a mortality benefitters? Defaults to FALSE.}

\item{mortality_trajectory_shape}{named character vector, valid values are: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}. Can differ across arms.}

\item{mortality_dampening}{named numeric vector, dampening effect on HRQoL at ICU discharge in
patients who die before end of follow-up}

\item{sampling_frequency}{named int vector, span between HRQoL sampling from patients, in each arm.}

\item{n_digits}{int, the number of digits of HRQoL values. More digits will yield greater
precision but also cause longer run-times.}
}
\value{
A three-element list with trajectories for each of the three assumptions as to when
sampling can start.
}
\description{
Construct the full (primary) HRQoL trajectory for a patient
}
\keyword{internal}
