% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_patient_trajectory.R
\name{construct_patient_trajectory}
\alias{construct_patient_trajectory}
\title{Construct the full (primary) HRQoL trajectory for a patient}
\usage{
construct_patient_trajectory(
  t_icu_discharge,
  t_death,
  first_hrqol_patient,
  acceleration_hrqol,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  is_mortality_benefitter,
  mortality_trajectory_shape,
  mortality_dampening,
  sampling_frequency,
  valid_hrqol_range
)
}
\arguments{
\item{t_icu_discharge}{scalar, time of ICU discharge}

\item{t_death}{numeric scalar, time of death. NA represent administrative censoring (i.e., the
patient was still alive at end of follow-up)}

\item{first_hrqol_patient}{single numeric value, the HRQoL value at ICU discharge of this patient}

\item{acceleration_hrqol}{scalar in \verb{[0, Inf)}, acceleration of HRQoL improvement compared to baseline
trajectory}

\item{first_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{is_mortality_benefitter}{single logical value, is patient a mortality benefitters? Defaults to \code{FALSE}.}

\item{sampling_frequency}{integer (single value or named vector), span between HRQoL sampling from patients}

\item{valid_hrqol_range}{two-element numeric vector, the lower and upper
bounds of valid HRQoL values. The default (\code{c(-0.757, 1.0)}) corresponds to
the Danish EQ-5D-5L index values.}
}
\value{
A three-element list with trajectories for each of the three assumptions as to when
sampling can start.
}
\description{
Construct the full (primary) HRQoL trajectory for a patient
}
\keyword{internal}
