% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_xout}
\alias{create_xout}
\title{Create xout vector for approx()}
\usage{
create_xout(start, end, by)
}
\arguments{
\item{start, end, by}{integers of length 1}
}
\description{
Importantly, there are no checks so if, e.g., \code{end = NA}, the functions
happily spits out a vector with some 2^29 elements. This lack of check is
intentional to keep this (internal) function fast.
}
\keyword{internal}
