% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helper.R
\name{estimation_helper}
\alias{estimation_helper}
\title{Helper function to estimate outcomes in individual patients}
\usage{
estimation_helper(
  n_patients,
  arm,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  inter_patient_noise_sd,
  acceleration_hrqol,
  prop_mortality_benefitters,
  mortality_trajectory_shape,
  mortality_dampening,
  mortality_rng,
  sampling_frequency = NULL,
  n_digits = 2,
  valid_hrqol_range = c(-0.757, 1)
)
}
\arguments{
\item{n_patients}{int, the number of patients for whom to estimate outcomes}

\item{arm}{character scalar}

\item{first_hrqol_arm}{scalar, HRQoL at ICU discharge (at arm level)}

\item{final_hrqol_arm}{scalar HRQoL at end of follow-up (at arm level)}

\item{acceleration_hrqol}{scalar, acceleration of HRQoL improvement}

\item{prop_mortality_benefitters}{named numeric vector \verb{[0, 1]}, the proportion of patients in
each arm who are so-called mortality benefitters.}

\item{mortality_trajectory_shape}{named character vector, valid values are: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}. Can differ across arms.}

\item{mortality_dampening}{named numeric vector, dampening effect on HRQoL at ICU discharge in
patients who die before end of follow-up}

\item{mortality_rng}{function, a function that produces times of death (only one parameter which should be the number of values to sample)}

\item{sampling_frequency}{int, for span between samplings from patients by the elements.}

\item{n_digits}{int, the number of digits of HRQoL values. More digits will yield greater
precision but also cause longer run-times.}

\item{valid_hrqol_range}{two-element numeric vector, the lower and upper bounds of valid HRQoL
values. The default (\code{c(-0.757, 1.0)}) corresponds to the Danish EQ-5D-5L index values.}
}
\value{
Data.table with outcome results for all patients. NB! There are no trial_id indicator as
this is not required in all uses of this helper function.
}
\description{
Helper function to estimate outcomes in individual patients
}
\keyword{internal}
