% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_mean}
\alias{fast_mean}
\title{Mean value without the housekeeping}
\usage{
fast_mean(x)
}
\arguments{
\item{x}{numeric vector, crucially this must be a numeric vector; e.g.
logical vectors must be wrapped in \code{as.double()}}
}
\value{
single numeric vector with the mean.
}
\description{
R's built-it \code{mean} function has a lot of overhead which is redundant for the
internal workings of \code{hrqolr}.
}
\keyword{internal}
