% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_decay_halflife.R
\name{find_decay_halflife}
\alias{find_decay_halflife}
\title{Find appropriate half-life of exponential decay}
\usage{
find_decay_halflife(
  t_diff = 180 - 1,
  min_halflife = 0,
  max_halflife = 180,
  res = 0.1,
  tol = 0.05
)
}
\arguments{
\item{t_diff}{int, the span in days that the decay should span}

\item{min_halflife, max_halflife}{integers, shortest and longets half lives to consider}

\item{res}{scalar numeric, grid size in half-life values to be checked}

\item{tol}{scalar, the proportion of the original HRQoL that corresponds to zero (the default
value corresponds to considering \verb{<5\%} of the starting HRQoL to be the same as \code{0})}
}
\value{
The best half-life
}
\description{
The notion of this function is to find the largest half-life that yields a almost-zero decay
factor (as per \code{tol} argument) at \code{grid_end}.
}
\keyword{internal}
