% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mortality_funs.R
\name{generate_mortality_funs}
\alias{generate_mortality_funs}
\title{Generate density, probability, quantile and random generation functions for (scaled) empirical
mortality distribution}
\usage{
generate_mortality_funs(cum_mortality = NULL, censoring_value = NA)
}
\arguments{
\item{cum_mortality}{scalar in [0, 1], the cumulative mortality at end of follow-up (= time of
censoring)}

\item{censoring_value}{the value assigned when a patient is censored, default is `NA`}
}
\value{
A list with the four functions, named `d`, `p`, `q` and `r` to follow R conventions (see,
  e.g., `?rnorm`)
}
\description{
The cumulative mortality functions is modelled over that of the CLASSIC trial, but can be scaled
to achieve a desired mortality at end of follow-up
}
