% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_matrix}
\alias{na_matrix}
\title{Create "empty" observation tibble}
\usage{
na_matrix(t)
}
\arguments{
\item{t}{scalar}
}
\description{
Useful for patients for whom no data are every observed.
}
\keyword{internal}
