% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_matrix}
\alias{na_matrix}
\title{Create "empty" observation tibble}
\usage{
na_matrix(index_hrqol, t)
}
\arguments{
\item{index_hrqol}{scalar, HRQoL at time of enrolment (= index)}

\item{t}{scalar}
}
\description{
Useful for patients for whom no data are every observed.
}
\keyword{internal}
