% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.hrqolr_trajectories}
\alias{plot.hrqolr_trial}
\title{Plotting methods for HRQoL results}
\usage{
\method{plot}{hrqolr_trajectories}(
  x,
  which = "both",
  arm_aes = NULL,
  patient_aes = list(alpha = 0.2),
  ribbon_aes = list(alpha = 0.2),
  ribbon_percentiles = c(0.25, 0.75),
  ...
)

\method{plot}{hrqolr_trial}(x, analysis = "all", ecdf = TRUE, ...)
}
\arguments{
\item{x}{object to plot, see \emph{Details}.}

\item{which}{character, one of \code{"both"} (arm and patient trajectories, the default), \code{"arm"}
(only arm trajectories), \code{"patient"} (only patient trajectories) and \code{"summarise"} (arm
trajectories with ribbons summarising the patient trajectories).}

\item{arm_aes, patient_aes, ribbon_aes}{lists with additional settings for the arm, patients and
ribbon layers.}

\item{ribbon_percentiles}{scalar of two elements defining the lower and upper percentiles of band
drawn to summarise patient trajectories}

\item{...}{passed on to methods}

\item{analysis}{character, \code{"all"} (include non-survivors and set their HRQoL to 0) or \code{"survivors"}.}

\item{ecdf}{logical, whether to plot the empiricial cumulative distribution (default) or not}
}
\value{
A ggplot2 object.

A \code{ggplot}.
}
\description{
Plotting methods for HRQoL results

Plot method for example trajectories

Plot results from single trial
}
\section{Methods (by class)}{
\itemize{
\item \code{plot(hrqolr_trajectories)}: Example trajectories

\item \code{plot(hrqolr_trial)}: Single trial

}}
