% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.hrqolr_trajectories}
\alias{print.hrqolr_scenario_validation_results}
\alias{print.hrqolr_scenario}
\alias{print.hrqolr_comparisons}
\alias{print.hrqolr_summary_stats}
\alias{print.hrqolr_bytes}
\title{Print methods for HRQoL results}
\usage{
\method{print}{hrqolr_trajectories}(x, prop_decimals = 3, ...)

\method{print}{hrqolr_scenario_validation_results}(x, ...)

\method{print}{hrqolr_scenario}(x, ...)

\method{print}{hrqolr_comparisons}(x, decimals = 3, ...)

\method{print}{hrqolr_summary_stats}(x, decimals = 3, ...)

\method{print}{hrqolr_bytes}(x, digits = 3, ...)
}
\arguments{
\item{x}{the object to be printed.}

\item{prop_decimals}{single integer (default is \code{3}), the number of digits used
when printing proportions.}

\item{...}{not used}

\item{decimals}{single integer (default is \code{3}), the number of digits}

\item{digits}{a non-null value for \code{digits} specifies the minimum
    number of significant digits to be printed in values.  The default,
    \code{NULL}, uses \code{\link[base]{getOption}("digits")}.  (For the
    interpretation for complex numbers see \code{\link[base]{signif}}.)
    Non-integer values will be rounded down, and only values
    greater than or equal to 1 and no greater than 22 are accepted.
  }
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints contents of the first input \code{x} in a human-friendly way, see
\strong{Details} for more information.

Logic from \code{pryr:::print.pryr_bytes} but simplified and included here to minimise dependencies.
}
\details{
The behaviour depends on the class of \code{x}:
}
