% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.hrqolr_examples}
\title{Print methods for HRQoL results}
\usage{
\method{print}{hrqolr_examples}(x, prop_digits = 3, ...)
}
\arguments{
\item{x}{object to print, see \strong{Details}.}

\item{prop_digits}{single integer (default is \code{3}), the number of digits used
when printing proportions.}

\item{...}{passed on to methods}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints contents of the first input \code{x} in a human-friendly way, see
\strong{Details} for more information.
}
\details{
The behaviour depends on the class of \code{x}:
}
\section{Methods (by class)}{
\itemize{
\item \code{print(hrqolr_examples)}: Example trajectories

}}
