% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_example_trajectories.R
\name{sample_example_trajectories}
\alias{sample_example_trajectories}
\title{Sample example trajectories}
\usage{
sample_example_trajectories(
  n_trajectories_per_arm = 100,
  start_hrqol_ctrl = 0.1,
  final_hrqol_ctrl = 0.75,
  relative_improvement_hrqol_actv = 0.1,
  sampling_frequency = 14L,
  acceleration_hrqol_actv = 0.1,
  mortality_ctrl = 0.4,
  relative_mortality_reduction_actv = 0.05,
  mortality_dampening = 0.7,
  mortality_trajectory_shape = "exp_decay",
  prop_mortality_benefitters_actv = 0.1,
  n_digits = 2,
  seed = NULL,
  ...
)
}
\arguments{
\item{start_hrqol_ctrl, final_hrqol_ctrl}{numeric scalars, the HRQoL at ICU discharge and end of
follow-up in the control arm, respectively}

\item{relative_improvement_hrqol_actv}{scalar numeric, e.g. \code{0.10} means that the HRQoL at end of
follow-up in the active arm will be increased by \verb{10\%} compared to the control arm}

\item{sampling_frequency}{int, for span between samplings from patients}

\item{acceleration_hrqol_actv}{scalar, relative acceleration of HRQoL improvement in the active
arm}

\item{mortality_ctrl}{numeric scalar, the mortality in the control group at end of follow-up}

\item{relative_mortality_reduction_actv}{numeric scalar, e.g. \code{0.10} means that the mortality in
the active arm is \verb{90\%} of that in the control arm}

\item{mortality_dampening}{scalar, dampening effect of HRQoL at ICU discharge in patients who die
before end of follow-up}

\item{mortality_trajectory_shape}{string, should be any of the following four: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}}

\item{prop_mortality_benefitters_actv}{scalar numeric in \verb{[0, 1]}, the proportion of patients in
the active arm who are so-called mortality benefitters.}

\item{n_digits}{int, the number of digits of HRQoL values}

\item{seed}{int, optional seed for reproducible pseudo-random number generation. Defaults to a
deterministic value based on the arguments given (ensuring reproducibility by default).}

\item{...}{not used}
}
\value{
An object of class \code{hrqolr_examples} with trajectories in data.tables. The function will
always return arm-level trajectories. If \code{n_trajectories_per_arm > 0}, it will also return
data.table with patient trajectories, with a \code{id} column denoting patients. with
}
\description{
Sample a desired number of trajectories at arm leven and for individual patients.
}
