% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_t_icu_discharge}
\alias{sample_t_icu_discharge}
\title{Sample time to ICU discharge}
\usage{
sample_t_icu_discharge(n)
}
\arguments{
\item{n}{single integer, number of times to ICU discharge to sample.}
}
\value{
numeric vector of length \code{n} with sampled times.
}
\description{
Helper function.
}
\keyword{internal}
