% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trials.R
\name{simulate_trials}
\alias{simulate_trials}
\title{Title}
\usage{
simulate_trials(
  n_trials = 100000L,
  n_patients_per_arm = 100L,
  start_hrqol_ctrl = 0.1,
  final_hrqol_ctrl = 0.75,
  relative_improvement_hrqol_actv = 0.1,
  sampling_frequency = 14L,
  acceleration_hrqol_actv = 0.1,
  mortality_ctrl = 0.3,
  relative_mortality_reduction_actv = 0.05,
  mortality_dampening = 0.7,
  mortality_trajectory = "exp_decay",
  prop_mortality_benefitters_actv = 0.1,
  n_digits = 2,
  seed = NULL,
  n_patients_ground_truth = 100000L
)
}
\arguments{
\item{n_trials}{int scalar or vector. If vector, simulations will be run in batches of size given by the elements.}

\item{n_patients_per_arm}{int scalar}

\item{start_hrqol_ctrl}{double}

\item{final_hrqol_ctrl}{double}

\item{relative_improvement_hrqol_actv}{double}

\item{sampling_frequency}{int}

\item{seed}{}
}
\value{
A string indicating where results were saved (if to a database) or a data.table (otherwise)
}
\description{
Title
}
