% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trials.R
\name{simulate_trials}
\alias{simulate_trials}
\alias{simulate_trials.hrqolr_scenario}
\alias{simulate_trials.default}
\title{Simulate trials}
\usage{
simulate_trials(...)

\method{simulate_trials}{hrqolr_scenario}(
  scenario,
  n_trials = 100,
  n_patients_ground_truth = 1000,
  n_example_trajectories_per_arm = 50,
  test_fun = welch_t_test,
  verbose = TRUE,
  n_digits = 2,
  seed = NULL,
  valid_hrqol_range = c(-0.757, 1),
  alpha = 0.05,
  ...
)

\method{simulate_trials}{default}(
  arms,
  n_patients_per_arm,
  sampling_frequency,
  index_hrqol,
  first_hrqol,
  final_hrqol,
  acceleration_hrqol,
  mortality,
  mortality_dampening,
  mortality_trajectory_shape,
  prop_mortality_benefitters,
  n_trials = 100,
  n_patients_ground_truth = 1000,
  n_example_trajectories_per_arm = 50,
  test_fun = welch_t_test,
  verbose = TRUE,
  n_digits = 2,
  seed = NULL,
  valid_hrqol_range = c(-0.757, 1),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{...}{not used}

\item{scenario}{object of class 'hrqolr_scenario', the output of \link{setup_scenario}}

\item{n_trials}{int scalar or vector. If vector, simulations will be run in batches of size given}

\item{n_patients_ground_truth}{int, how many patients (per arm) to use when estimating the ground
truth}

\item{n_example_trajectories_per_arm}{int, the number of example trajectories to include in the
returned object}

\item{test_fun}{function used to compare estimates. The default is \code{welch_t_test} (built into
\code{hrqolr}). Note that the function should handle the number of arms provided. See \strong{Details}
below.}

\item{verbose}{logical, should the function give progress timestamped updates? Default: \code{TRUE}}

\item{n_digits}{int, the number of digits of HRQoL values. More digits will yield greater
precision but also cause longer run-times.}

\item{seed}{int, optional seed for reproducible pseudo-random number generation. Defaults to a
deterministic value based on the arguments given (ensuring reproducibility by default).}

\item{valid_hrqol_range}{two-element numeric vector, the lower and upper bounds of valid HRQoL
values. The default (\code{c(-0.757, 1.0)}) corresponds to the Danish EQ-5D-5L index values.}

\item{alpha}{scalar in \verb{[0, 1]}, the desired type 1 error rate used when comparing HRQoL in the
arms.}

\item{arms}{character vector with the names of the arms. Must match the names of named vectors
below.}

\item{n_patients_per_arm}{named int vector, number of patient in each arm}

\item{sampling_frequency}{named int vector, span between HRQoL sampling from patients, in each arm.}

\item{index_hrqol}{named numeric vector, the HRQoL at index (= enrolment)}

\item{first_hrqol}{named numeric vector, the HRQoL at ICU discharge in each arm}

\item{final_hrqol}{named numeric vector, the HRQoL at end of follow-up in each arm}

\item{acceleration_hrqol}{named numeric vector, relative acceleration of HRQoL improvement in
each arm}

\item{mortality}{named numeric vector in \verb{[0, 1]}, the mortality in at end of follow-up, in each
arm}

\item{mortality_dampening}{named numeric vector, dampening effect on HRQoL at ICU discharge in
patients who die before end of follow-up}

\item{mortality_trajectory_shape}{named character vector, valid values are: \code{"exp_decay"}
(default), \code{"linear"}, \code{"constant"}, \code{"reflected_exp_decay"}. Can differ across arms.}

\item{prop_mortality_benefitters}{named numeric vector \verb{[0, 1]}, the proportion of patients in
each arm who are so-called mortality benefitters.}
}
\value{
An object of class \code{hrqolr_results}, which is a specialised list with four elements:
summary statistics for each arm, comparisons (incl. performance metrics), the seed and the
elapsed time.
}
\description{
This is the key user-facing function for simulating trials.
}
\details{
\itemize{
\item \code{test_fun}: [pending]
}
}
